requirements.txt
django-admin-interface
django-widget-tweaks
qrcode
django-dynamic-db-router
djangorestframework
opencv-python
